refactored file structor
