+++ front-end/.env
@@ -1,0 +1,1 @@
+REACT_APP_DB_URL=http://localhost:7777
